Split fixtures into docker/slurm
