+++ tests/conftest.py
@@ -64,6 +64,7 @@ from .fixtures_server import *  # noqa F403
 from .fixtures_server_v1 import *  # noqa F403
 from .fixtures_server_v2 import *  # noqa F403
 from .fixtures_tasks_v1 import *  # noqa F403
-from .fixtures_slurm import *  # noqa F403
 from .fixtures_tasks_v2 import *  # noqa F403
+from .fixtures_docker import *  # noqa F403
+from .fixtures_slurm import *  # noqa F403
 from .fixtures_commands import *  # noqa F403
